% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcensus.R
\name{qcensus}
\alias{qcensus}
\title{qcensus}
\usage{
qcensus(yearz, agez)
}
\arguments{
\item{yearz}{int or vector of years to calculate 2016 - 2060}

\item{agez}{int or vector of ages to calculate 0 - 100}
}
\description{
Calculate number of people in the US population for a given range of years and ages
}
\examples{
qcensus(yearz = c(2016, 2017), agez = c(0, 100))
}
